+++ src/RcppExports.cpp
@@ -62,7 +62,7 @@ BEGIN_RCPP
 END_RCPP
 }
 // phi_update
-Rcpp::List phi_update(double phi_old, double sigma2_theta, arma::vec theta, Rcpp::List temporal_corr_info, double a_phi, double b_phi, double metrop_var_phi_trans, double acctot_phi_trans);
+Rcpp::List phi_update(double phi_old, double sigma2_theta, arma::vec theta, Rcpp::List temporal_corr_info, double a_phi, double b_phi, double metrop_var_phi_trans, int acctot_phi_trans);
 RcppExport SEXP _GPCW_phi_update(SEXP phi_oldSEXP, SEXP sigma2_thetaSEXP, SEXP thetaSEXP, SEXP temporal_corr_infoSEXP, SEXP a_phiSEXP, SEXP b_phiSEXP, SEXP metrop_var_phi_transSEXP, SEXP acctot_phi_transSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -74,7 +74,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type a_phi(a_phiSEXP);
     Rcpp::traits::input_parameter< double >::type b_phi(b_phiSEXP);
     Rcpp::traits::input_parameter< double >::type metrop_var_phi_trans(metrop_var_phi_transSEXP);
-    Rcpp::traits::input_parameter< double >::type acctot_phi_trans(acctot_phi_transSEXP);
+    Rcpp::traits::input_parameter< int >::type acctot_phi_trans(acctot_phi_transSEXP);
     rcpp_result_gen = Rcpp::wrap(phi_update(phi_old, sigma2_theta, theta, temporal_corr_info, a_phi, b_phi, metrop_var_phi_trans, acctot_phi_trans));
     return rcpp_result_gen;
 END_RCPP
